Explanation of how the program works.
main.cpp is the program that takes an algorithm (can be source code or psudo-code) as inputFile and calculate the Asymptotic RunTime Complexity of that algorithm by checking the number of tiems a loop ( for or while) has counted in a single run.

for more details, check main.cpp file and read the comments.